ABC174 B, refine

+++ exercise/src/main.rs
@@ -1,12 +1,12 @@
 fn main() {
     proconio::input! {
         n: usize,
-        d: usize,
+        d: f64,
         v: [(f64, f64); n],
     }
     let mut ans = 0;
     for &(x, y) in &v {
-        if x.hypot(y) <= d as f64 {
+        if x.hypot(y) <= d {
             ans += 1;
         }
     }
